Create navigation bar framework
