+++ package.json
@@ -10,6 +10,8 @@
     "canned": "^0.3.13",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
+    "react-router": "^6.2.1",
     "react-scripts": "5.0.0",
     "sass": "^1.47.0",
     "web-vitals": "^2.1.2"
