+++ server/banners/index.get.json
@@ -1,34 +1,34 @@
 [
   {
-    "bannerImageUrl": "/static/images/offers/offer1.jpg",
+    "bannerImageUrl": "/assets/images/offers/offer1.jpg",
     "bannerImageAlt": "Independence Day Deal - 25% off on shampoo",
     "isActive": true,
     "order": 1,
     "id": "5b6c38156cb7d770b7010ccc"
   },
   {
-    "bannerImageUrl": "/static/images/offers/offer2.jpg",
+    "bannerImageUrl": "/assets/images/offers/offer2.jpg",
     "bannerImageAlt": "Independence Day Deal - Rs120 off on surf",
     "isActive": true,
     "order": 2,
     "id": "5b6c38336cb7d770b7010ccd"
   },
   {
-    "bannerImageUrl": "/static/images/offers/offer3.jpg",
+    "bannerImageUrl": "/assets/images/offers/offer3.jpg",
     "bannerImageAlt": "Independence Day Deal - Rs99 off on domex",
     "isActive": true,
     "order": 3,
     "id": "5b6c38456cb7d770b7010cce"
   },
   {
-    "bannerImageUrl": "/static/images/offers/offer4.jpg",
+    "bannerImageUrl": "/assets/images/offers/offer4.jpg",
     "bannerImageAlt": "Independence Day Deal - Rs99 off on bodywash",
     "isActive": true,
     "order": 4,
     "id": "5b6c38576cb7d770b7010ccf"
   },
   {
-    "bannerImageUrl": "/static/images/offers/offer5.jpg",
+    "bannerImageUrl": "/assets/images/offers/offer5.jpg",
     "bannerImageAlt": "Independence Day Deal - Rs70 off on tea",
     "isActive": true,
     "order": 5,
