+++ server/categories/index.get.json
@@ -5,7 +5,7 @@
     "description": "Our beverage department will ensure your fridge is always fully stocked. Shop for soda, juice, beer and more. ",
     "enabled": true,
     "order": 3,
-    "imageUrl": "/static/images/category/beverages.png",
+    "imageUrl": "/assets/images/category/beverages.png",
     "id": "5b675e5e5936635728f9fc30"
   },
   {
@@ -14,7 +14,7 @@
     "description": "The best cupcakes, cookies, cakes, pies, cheesecakes, fresh bread, biscotti, muffins, bagels, fresh coffee, milk and more.",
     "enabled": true,
     "order": 2,
-    "imageUrl": "/static/images/category/bakery.png",
+    "imageUrl": "/assets/images/category/bakery.png",
     "id": "5b6899123d1a866534f516de"
   },
   {
@@ -23,7 +23,7 @@
     "description": "Buy beauty and personal care products online in India at best prices.",
     "enabled": true,
     "order": 4,
-    "imageUrl": "/static/images/category/beauty.png",
+    "imageUrl": "/assets/images/category/beauty.png",
     "id": "5b68994e3d1a866534f516df"
   },
   {
@@ -32,7 +32,7 @@
     "description": "Shop online for Baby Products, Diapers, Skin Care Products,etc.",
     "enabled": true,
     "order": 5,
-    "imageUrl": "/static/images/category/baby.png",
+    "imageUrl": "/assets/images/category/baby.png",
     "id": "5b6899683d1a866534f516e0"
   },
   {
@@ -49,7 +49,7 @@
     "description": "A variety of fresh fruits and vegetables.",
     "enabled": true,
     "order": 1,
-    "imageUrl": "/static/images/category/fruits.png",
+    "imageUrl": "/assets/images/category/fruits.png",
     "id": "5b6899953d1a866534f516e2"
   }
 ]
