+++ src/assets/images/category/fruits.png
