+++ src/common/_mixins.scss
@@ -1,0 +1,5 @@
+@mixin centralize {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
