+++ src/common/_variables.scss
@@ -1,3 +1,3 @@
-$primary: rgb(68, 171, 231);
-$gray: rgb(49, 9, 9);
-$accent: rgb(177, 226, 85);
+$primary: #bf2957;
+$gray: #033f46;
+$accent: #79d7e1;
